Don't use CLLADDR() on FreeBSD.
This macro was removed in FreeBSD 15.  There is no reason to use it
in this code on any version of FreeBSD.  The constness is provided
by the const struct sockaddr_dl * pointer.  The base type of char
of sdl_data gives us all needed address arithmetics, no need to use
the relic c_caddr_t.